+++ EMS/src/TimeSheet.Modules.EmploymentManagement.Api/Authentication/ILoginRateLimitKeyProvider.cs
@@ -1,0 +1,6 @@
+namespace TimeSheet.Modules.EmploymentManagement.Api.Authentication;
+
+public interface ILoginRateLimitKeyProvider
+{
+    string GetPartitionKey(HttpContext httpContext);
+}